Code changes trying to improve the change genre
We do not need Call returns on providers because we do not cancel individual requests.
Removed not neccesary changed param on success callbacks.

Simplified PagerAdapter for media, instantiateItem is not neccesary , the (super) fragment manager do the job.

Improved code for MediaProvider.Callback on MediaListFragment.

Fixed deprecation warning on LoadingDetailDialogFragment.

Tv module changes to adapt it to below changes.